fix(audit): split census out of storage.rs under the 1000-line cap
Move count/list_sessions off the storage facade. No comment stripping.

Closes #1596
